public SecurityMockMvcRequestPostProcessors and TestSecurityContextRepository for reusability
